doc: Capture settings to enable Memory Manager in the install guide
Following-up on the documentation update made in https://github.com/k8stopologyawareschedwg/documentation/pull/5,
we ensure that the install guide captures kubelet settings pertaining
to memory manager.

Memory support was enabled in NFD-topology-updater (https://github.com/kubernetes-sigs/node-feature-discovery/pull/593)
and this update reflects that we support memory and its
accounting on a per NUMA basis in addition to CPU and devices.

This patch:
- Provides examples on how to configure system with memory manager
- Points to upstream documentation if more info is needed
- Adds an example kind config for reference

Signed-off-by: Swati Sehgal <swati.tcd@gmail.com>